+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Implement mapping profile deletion feature (UIDATIMP-229)
 * Make mapping profile name a hotlink in action profile details (UIDATIMP-214)
 * Add tags accordion for match profiles (UIDATIMP-18)
-
+* Create TreeView component (UIDATIMP-243)
 
 ## [1.4.0](https://github.com/folio-org/ui-data-import/tree/v1.4.0) (2019-08-02)
 * Create data import settings page's 3rd pane for action profiles (UIDATIMP-8)